apps/training-service/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riskops-training-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.10.4",
  "pydantic-settings==2.7.0",
  "psycopg2-binary==2.9.9",
  "sqlalchemy==2.0.36",
  "pandas==2.2.3",
  "numpy==2.2.1",
  "arch==7.2.0",
  "scipy==1.14.1",
  "mlflow==2.19.0",
  "boto3==1.35.86",
  "kafka-python==2.0.2",
  "matplotlib==3.9.4",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["training_service*"]
